Extract combinePageInterpretationsForSheet to libs/ballot-interpreter
Behavior-preserving move so both VxScan and VxCentralScan can consume the
same sheet-interpretation combiner. The function and its unit tests move
verbatim; VxScan's interpret.ts now imports from the lib.

Adds a single-page BMD unit test pair (mirrors the existing multi-page
BMD tests) so the lib reaches 100% coverage on the combiner — those two
cases were previously only exercised by VxScan's integration tests.
