+++ libs/ballot-interpreter/src/index.ts
@@ -1,4 +1,5 @@
 /* istanbul ignore file - @preserve */
+export * from './combine_page_interpretations_for_sheet';
 export * from './interpret';
 export * from './bubble-ballot-ts';
 export * from './save_images';
